refactor(agent): split macos/collector.rs into submodules
Convert apps/agent/src-tauri/src/macos/collector.rs (1,673 lines) into a
macos/collector/ directory broken into focused submodules:

- mod.rs (216): public start_status_collector and the long-running poll
  loop that drives status probes, refreshes, and network-extension reloads
- tool.rs (219): ToolInvocation enum, helper-tool execution with env
  scrubbing, run_json_tool decoder, and helper-tool env-var constants
- resolve.rs (162): tool resolution via env overrides, Tauri resource
  bundle, or local Swift package build outputs
- merge.rs (259): combined-status merge, provider enrichment, runtime-
  snapshot staleness checks, and activation-state derivation
- samples.rs (76): wire-format Deserialize types for helper output
- reload.rs (79): NetworkExtension reload helper invocation + response
  validation
- paths.rs (22): default endpoint-security / network-extension snapshot
  and egress-policy file locations
- tests.rs (762): all existing unit tests, unmodified in behavior

No external API change; macos/mod.rs still re-exports start_status_collector.
All production files are well under the 600-line cap.
